+++ backlib/internal/utils/alias.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from enum import Enum, auto
 from typing import Any, TypeVar
 
-from backlib.py313.internal.utils.platform import (
+from backlib.internal.utils.platform import (
     is_darwin,
     is_freebsd,
     is_linux,